dto layer

+++ account-service/src/main/kotlin/com/transferapp/accountservice/adapter/in/dto/AccountDTORequest.kt
@@ -5,5 +5,5 @@ import javax.validation.constraints.NotBlank
 data class AccountDTORequest(
     @field:NotBlank
     val customerDocumentId: String,
-    val number: Long
+    val number: Int
     )
